Moved tests into the correct place.
