+++ kieker-analysis/test/kieker/analysis/architecture/dependency/package-info.java
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package kieker.analysis.junit.plugin.filter.select;
+package kieker.analysis.architecture.dependency;
