+++ kieker-analysis/test/kieker/analysis/architecture/recovery/package-info.java
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package kieker.analysis.dependencygraphs;
+package kieker.analysis.architecture.recovery;
