+++ kieker-analysis/test/kieker/analysis/generic/time/package-info.java
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package kieker.analysis.signature;
+package kieker.analysis.generic.time;
